#35 use projected cols as literals if no table is given
If no table to select from is given to the compiler in a select core, the compiler will use the projected column names as values in the returned command, and will project upon those values with the aliases that were given in the original projection.
This allows for statements like `SELECT NOW() AS date`, which will compile to a projection `column1 -> date` over `VALUES (NOW())`.