dbfs: implement dbfs validation
Upon loading a database, the file structure as well as the content
of the tables.info file will be validated.
If validation fails, the structure cannot be loaded as a (*DBFS).

Also changed the content of the tables.info file to now also hold the
count of tables in the database.